requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
pymysql==1.1.1
openai==1.57.2
aiofiles==24.1.0
Pillow==11.0.0
python-multipart==0.0.20
mysql-connector-python==9.2.0
python-jose==3.4.0